Refactor: GeocodingService를 LocationService로 추상화
- 테스트용 LocationService 생성
