+++ src/main/java/com/stoury/service/location/LocationService.java
@@ -1,0 +1,7 @@
+package com.stoury.service.location;
+
+import com.stoury.dto.LocationResponse;
+
+public interface LocationService {
+    LocationResponse getLocationFrom(double latitude, double longitude);
+}
