+++ src/test/java/com/stoury/service/TestLocationService.java
@@ -1,11 +1,14 @@
 package com.stoury.service;
 
 import com.stoury.dto.LocationResponse;
+import com.stoury.service.location.LocationService;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
 @Service
-public class GeocodingService {
-    // TODO: 지오코딩 api를 이용해 구현 필요
+@Profile("test")
+public class TestLocationService implements LocationService {
+    @Override
     public LocationResponse getLocationFrom(double latitude, double longitude) {
         return new LocationResponse("default_city", "default_country");
     }
